Use nodemon in npm scripts to auto-reload server
[nodemon](https://github.com/remy/nodemon) can be used to restart the
dev server when server-related files are changed. An `npm start` script
is exposed to run the server with nodemon, so that it will restart on
any changes.

[Grunt's "watch" task](https://github.com/gruntjs/grunt-contrib-watch)
will lint and test the application whenever a change is detected.

[node-inspector](https://github.com/node-inspector/node-inspector) is
supported for live-debugging with the Chrome developer tools. An
`npm run debug` script is included to start a debugging instance of
the dev server; this will also be restarted with nodemon if a change
is detected.

CSS files, templates and client-side JS are excluded from the build
because Express reloads all of them on each request, and Stylus files
are handled through the stylus middleware.